requirements.txt
webdavclient3
langchain
faiss-cpu
pypdf2
openai
python-dotenv
langchain_community
chromadb
sentence-transformers
Flask 
flask-cors
